# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для LMS платформы Cap Education.",
        "title": "Cap Education LMS - API",
        "contact": {},
        "version": "1.0"
    },
    "host": "capapi.grtsq.ru",
    "basePath": "/",
    "paths": {
...
        "/courses/{id}": {
            "get": {
                "description": "Получить полную структуру курса (модули, уроки) с отметками о прохождении.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "УЧЕНИК: Страница курса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID курса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StudentCourseView"
                        }
                    }
                }
            }
        },
...
        "/lessons/{id}": {
            "get": {
                "description": "Получить контент конкретного урока (видео, текст).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "УЧЕНИК: Просмотр урока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID урока",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StudentLessonDetail"
                        }
                    }
                }
            }
        },
        "/lessons/{id}/assignment": {
            "post": {
                "description": "Отправка текстового ответа или файла.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "УЧЕНИК: Сдать домашнее задание",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID урока",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Текст ответа",
                        "name": "text_answer",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Файл с заданием",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lessons/{id}/complete": {
            "post": {
                "description": "Отметить урок как пройденный (для уроков без обязательного ДЗ).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "УЧЕНИК: Завершить урок (без ДЗ)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID урока",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/projects/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "USER: Детали проекта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Project"
                        }
                    }
                }
            }
        },
...
        "/teachers/{id}": {
            "get": {
                "description": "Получить полную инфо о преподавателе и его отзывы.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "УЧЕНИК: Детали преподавателя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID преподавателя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.TeacherPublicInfo"
                        }
                    }
                }
            }
        },
        "/teachers/{id}/reviews": {
            "post": {
                "description": "Добавить оценку и текстовый отзыв.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "УЧЕНИК: Оставить отзыв преподавателю",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID преподавателя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные отзыва",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/tests/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student-Learning"
                ],
                "summary": "USER: Детали теста",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Test ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Test"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.ContentBlock": {
            "type": "object",
            "properties": {
                "content": {},
                "type": {
                    "type": "string"
                }
            }
        },
        "domain.Course": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_homework": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_anti_copy_enabled": {
                    "type": "boolean"
                },
                "is_discord_mandatory": {
                    "type": "boolean"
                },
                "is_homework_mandatory": {
                    "type": "boolean"
                },
                "is_main": {
                    "type": "boolean"
                },
                "is_project_mandatory": {
                    "type": "boolean"
                },
                "is_test_mandatory": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/domain.CourseStatus"
                },
                "teacher_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.CourseStatus": {
            "type": "string",
            "enum": [
                "draft",
                "active",
                "archived"
            ],
            "x-enum-varnames": [
                "CourseStatusDraft",
                "CourseStatusActive",
                "CourseStatusArchived"
            ]
        },
...
        "domain.Lesson": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ContentBlock"
                    }
                },
                "content_text": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "duration_min": {
                    "type": "integer"
                },
                "has_homework": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_interactive": {
                    "type": "boolean"
                },
                "is_published": {
                    "type": "boolean"
                },
                "lesson_time": {
                    "type": "string"
                },
                "module_id": {
                    "type": "string"
                },
                "order_num": {
                    "type": "integer"
                },
                "presentation_url": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Project"
                    }
                },
                "teacher_id": {
                    "type": "string"
                },
                "tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Test"
                    }
                },
                "title": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "domain.LessonMaterial": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "string"
                },
                "public_url": {
                    "type": "string"
                },
                "s3_path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "domain.Project": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "string"
                },
                "max_score": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "domain.StudentCourseView": {
            "type": "object",
            "properties": {
                "course": {
                    "$ref": "#/definitions/domain.Course"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StudentModuleView"
                    }
                },
                "root_lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StudentLessonRef"
                    }
                },
                "root_projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Project"
                    }
                },
                "root_tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Test"
                    }
                }
            }
        },
        "domain.StudentLessonDetail": {
            "type": "object",
            "properties": {
                "assignment_status": {
                    "type": "string"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "lesson": {
                    "$ref": "#/definitions/domain.Lesson"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.LessonMaterial"
                    }
                },
                "next_lesson_id": {
                    "type": "string"
                },
                "previous_lesson_id": {
                    "type": "string"
                },
                "teacher_comment": {
                    "type": "string"
                }
            }
        },
        "domain.StudentLessonRef": {
            "type": "object",
            "properties": {
                "duration_min": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "is_locked": {
                    "type": "boolean"
                },
                "order_num": {
                    "type": "integer"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Project"
                    }
                },
                "tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Test"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.StudentModuleView": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StudentLessonRef"
                    }
                },
                "order_num": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "domain.TeacherPublicInfo": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "experience_years": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TeacherReview"
                    }
                },
                "schedule": {}
            }
        },
        "domain.TeacherReview": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "course_title": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "student_id": {
                    "type": "string"
                },
                "student_name": {
                    "type": "string"
                },
                "teacher_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.Test": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "string"
                },
                "passing_score": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "http.ReviewRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
...
    }
}